fix: gvm install idempotent

+++ run_onchange_before_05-dev-tools.sh
@@ -8,11 +8,12 @@ echo "Getting rust..."
 curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh -s -- -y
 
 # https://github.com/moovweb/gvm
-echo "Getting gvm..."
-rm -rf "$HOME/.gvm"
-bash < <(curl -s -S -L https://raw.githubusercontent.com/moovweb/gvm/master/binscripts/gvm-installer)
-source ~/.gvm/scripts/gvm
-gvm install $(gvm listall | grep -P '^\s*go\d' | sort -V | tail -n 1 ) -B
+if [[ ! -e ~/.gvm/scripts/gvm ]] ; then
+    echo "Getting gvm..."
+    bash < <(curl -s -S -L https://raw.githubusercontent.com/moovweb/gvm/master/binscripts/gvm-installer)
+    source ~/.gvm/scripts/gvm
+    gvm install $(gvm listall | grep -P '^\s*go\d' | sort -V | tail -n 1 ) -B
+fi
 
 # get asdf
 if [[ ! -d "$HOME/.asdf" ]] ; then
